Write to stdout instead of returning strings
I haven't done this so far because I thought that we need the strings
in order to properly test the functions. Now that I've found the
go-capturer package this turned out to be incorrect.

I wrote some tests in the previous commit using this package and they
work fine.

Writing to stdout has benefits:

- Text is immediately visible to the user. I.e. they can see progress
  instead of having to wait until the function is completely finished
- The code gets significantly simpler